pyproject.toml
[tool.poetry]
name = "combinator"
version = "0.1.0"
description = ""
authors = ["James <4589250+Naught0@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.3"
Flask-Cors = "^3.0.10"
peewee = "^3.14.10"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
pandas = "^1.4.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
